move general methods to utilities
